New: Put supplied bar numbers in the label field
This commit adds the ability to put non-consecutive bar numbers in the measure 'label' field.

Note: This only happens on measures with an explicit 'ExternalBarNumberAsString' property. Applications wishing to re-construct the bar numbers as supplied will have to implement their own counting, starting at a measure element with a label attribute defined.