Fixed: Lyric lines would get out of order
Previously, verse numbers were calculated by assuming that they were encoded sequentially. However, it turns out that Sibelius can encode the verse number in any order. This led to a bug where the verse numbers could be swapped.

This commit introduces a new way of tracking the lyric lines. It now parses the verse number out of the LyricItem StyleID and uses that for the verse number. This is more in line with how Sibelius differentiates the verses internally.

Also, this commit introduces a new function for _SplitString which splits a string into an array of characters if delimiter is false.