Fixed: Improved natural handling
This commit adds improved handling of explicitly-specified naturals. It will now produce accidentals for notes which have previously been modified in the measure.

However, it will not encode explicit naturals for notes that are a) in the key signature and b) have been previously altered as a natural in the measure. This is because Sibelius does not differentiate between "natural" and "no accidental", and will only provide a "cautionary" attribute on the first note that has been changed.

I'm going to say this closes the issue, but it's still not quite correct.

Fixes #20