New: Handle trills that are specified as SymbolItems
This commit adds support for trills that are specified as Symbols (not as lines).

If a trill is assigned to all voices this commit will raise a warning that MEI cannot handle this (how can you have a trill that is not in a voice?) and it will be assumed to be in voice (layer) 1.