Fixed: tupletSpan used instead of tuplet
This commit changes tuplet to tupletSpan. Since Sibelius makes it difficult to determine which notes are part of a tuplet ahead of time, tupletSpans are now used as milestone elements, instead of assuming tuplets containing notes.

This commit also fixes problems with the warnings dialog being null under certain conditions.